requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
yfinance==0.2.31
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
websockets==12.0 